# Conflicts:
#   backend/src/main/java/com/newsbig/sinmunmul/repository/ScrapRepositorySupport.java

+++ backend/src/main/java/com/newsbig/sinmunmul/repository/ScrapRepositorySupport.java
@@ -45,7 +45,7 @@ public class ScrapRepositorySupport {
 		return new PageImpl<>(scraps, pageable, total);
 	}
 	
-	public boolean checkScrap(int userSeq, int newsSeq) {
+	public boolean checkScrap(int userSeq, long newsSeq) {
 		try {
 			jpaQueryFactory.select(qScrap)
 				.from(qScrap)
@@ -59,7 +59,7 @@ public class ScrapRepositorySupport {
 		return true;
 	}
 	
-	public void deleteScrap(int userSeq, int newsSeq) {
+	public void deleteScrap(int userSeq, long newsSeq) {
 		List<Scrap> scraps = jpaQueryFactory.select(qScrap)
 				.from(qScrap)
 					.leftJoin(qUser).on(qUser.userSeq.eq(qScrap.user.userSeq))
